pupil_processor/requirements.txt
zmq 
msgpack==0.5.6
paho-mqtt>=1.6.1